// modules/server/src/domain/stats/types.ts
/** Ports the `DenpaMenStatBonus` record typedef from `denpa_men_stat_bonus.dart`. */
export interface DenpaMenStatBonus {
  hp: number;
  ap: number;
  attack: number;
  defense: number;
  speed: number;
  evasionRate: number;
}

export interface HeadShapeStatBonuses {
  hpBonus: number;
  apBonus: number;
  attackBonus: number;
  defenseBonus: number;
  speedBonus: number;
  evasionRateBonus: number;
}

export interface Correction {
  hpBonus: number;
  apBonus: number;
  attackBonus: number;
  defenseBonus: number;
  speedBonus: number;
  evasionRateBonus: number;
}